+++ packages/ice/src/commands/start.ts
@@ -11,7 +11,7 @@ import webpack from '@ice/bundles/compiled/webpack/index.js';
 import lodash from '@ice/bundles/compiled/lodash/index.js';
 import { getWebpackConfig } from '@ice/webpack-config';
 import type { ExtendsPluginAPI, ServerCompiler, GetAppConfig, GetRoutesConfig, GetDataloaderConfig } from '../types';
-import { ROUTER_MANIFEST, RUNTIME_TMP_DIR, WEB, IMPORT_META_TARGET, IMPORT_META_RENDERER } from '../constant.js';
+import { RUNTIME_TMP_DIR, WEB, IMPORT_META_TARGET, IMPORT_META_RENDERER } from '../constant.js';
 import webpackCompiler from '../service/webpackCompiler.js';
 import formatWebpackMessages from '../utils/formatWebpackMessages.js';
 import prepareURLs from '../utils/prepareURLs.js';
@@ -135,13 +135,12 @@ async function startDevServer({
       } else if (ssg) {
         renderMode = 'SSG';
       }
-      const routeManifestPath = path.join(rootDir, ROUTER_MANIFEST);
       // both ssr and ssg, should render the whole page in dev mode.
       const documentOnly = !ssr && !ssg;
 
       const serverRenderMiddleware = createRenderMiddleware({
         serverCompileTask,
-        routeManifestPath,
+        rootDir,
         documentOnly,
         renderMode,
         getAppConfig,