+++ packages/ice/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## v3.0.6
 
+- [feat] add suspense to exports
 - [fix] remove code when dead code is `ObjectProperties`
 - [fix] rebuild server entry when document changed
 - [fix] set ssr default value to `false`
@@ -14,7 +15,6 @@
 ## v3.0.4
 
 - [feat] support incremental compile for server bundle
-- [feat] add suspense to exports
 - [feat] refactor server bundle alias
 - [feat] support log debug info with different namespaces
 - [fix] support HMR when modules is used by dataLoader